pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "eva4-p2"
version = "0.1.0"
description = "All the course work part of EVA4 Phase 2 for the RS Group."
authors = [ "Ramjee Ganti <code@ramjeeganti.com>", "Roshan P <roshantac@gmail.com>", "Sujit Ojha <sujit.ojha@gmail.com>", "Srinivasan G <srinivasan.gsvasan@gmail.com>", "Dr P V Rajesh Kumar <drpvrajeshkumar@gmail.com>",]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
python-decouple = "^3.3"

[tool.poetry.dev-dependencies]
pelican = "^4.2.0"
rstcheck = "^3.3.1"
